requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
